// wadi-dsl/spike/langium-linking/src/generated/ast.ts
export interface Component extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Component';
    name: string;
    uses: Array<Use>;
}

export const Component = {
    $type: 'Component',
    name: 'name',
    uses: 'uses'
} as const;

export function isComponent(item: unknown): item is Component {
    return reflection.isInstance(item, Component.$type);
}

export interface Import extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Import';
    ns: string;
    path: string;
}

export const Import = {
    $type: 'Import',
    ns: 'ns',
    path: 'path'
} as const;

export function isImport(item: unknown): item is Import {
    return reflection.isInstance(item, Import.$type);
}

export interface Model extends langium.AstNode {
    readonly $type: 'Model';
    components: Array<Component>;
    imports: Array<Import>;
    uses: Array<Use>;
}

export const Model = {
    $type: 'Model',
    components: 'components',
    imports: 'imports',
    uses: 'uses'
} as const;

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model.$type);
}

export type QualifiedName = string;

export function isQualifiedName(item: unknown): item is QualifiedName {
    return typeof item === 'string';
}

export interface Use extends langium.AstNode {
    readonly $container: Component | Model;
    readonly $type: 'Use';
    target: langium.Reference<Component>;
}

export const Use = {
    $type: 'Use',
    target: 'target'
} as const;